Moved parse and serialize to MetaDataParser and MDSerializer, and simplified
