+++ src/main/java/org/jabref/cli/ArgumentProcessor.java
@@ -6,6 +6,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.EnumSet;
 import java.util.List;
 import java.util.Locale;
 import java.util.Optional;
@@ -603,8 +604,8 @@ public class ArgumentProcessor {
     private void importPreferences() {
         try {
             preferencesService.importPreferences(Path.of(cli.getPreferencesImport()));
-            Globals.entryTypesManager.addCustomOrModifiedTypes(preferencesService.getBibEntryTypes(BibDatabaseMode.BIBTEX),
-                    preferencesService.getBibEntryTypes(BibDatabaseMode.BIBLATEX));
+            EnumSet.allOf(BibDatabaseMode.class).forEach(mode ->
+                    Globals.entryTypesManager.addCustomOrModifiedTypes(preferencesService.getBibEntryTypes(mode), mode));
         } catch (JabRefException ex) {
             LOGGER.error("Cannot import preferences", ex);
         }
