+++ src/main/java/org/jabref/logic/exporter/MetaDataSerializer.java
@@ -6,11 +6,15 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
+import java.util.stream.Collectors;
 
 import org.jabref.logic.citationkeypattern.AbstractCitationKeyPattern;
 import org.jabref.logic.citationkeypattern.GlobalCitationKeyPattern;
 import org.jabref.logic.cleanup.FieldFormatterCleanups;
 import org.jabref.logic.util.OS;
+import org.jabref.model.entry.BibEntryType;
+import org.jabref.model.entry.field.BibField;
+import org.jabref.model.entry.field.FieldFactory;
 import org.jabref.model.entry.types.EntryType;
 import org.jabref.model.groups.GroupTreeNode;
 import org.jabref.model.metadata.ContentSelector;
@@ -131,5 +135,21 @@ public class MetaDataSerializer {
             stringBuilder.append(OS.NEWLINE);
         }
         return stringBuilder.toString();
+    }
+
+    public static String serializeCustomEntryTypes(BibEntryType entryType) {
+        StringBuilder builder = new StringBuilder();
+        builder.append(MetaData.ENTRYTYPE_FLAG);
+        builder.append(entryType.getType().getName());
+        builder.append(": req[");
+        builder.append(FieldFactory.serializeOrFieldsList(entryType.getRequiredFields()));
+        builder.append("] opt[");
+        builder.append(FieldFactory.serializeFieldsList(
+                entryType.getOptionalFields()
+                         .stream()
+                         .map(BibField::getField)
+                         .collect(Collectors.toList())));
+        builder.append("]");
+        return builder.toString();
     }
 }
